Further simplify C++ infrastructure.
The goal is for each implementation to consist of two files: a
run_foo.cpp file that does all the real logic, and a trivial foo.py
file. This is in contrast to the current situation of three nontrivial
C++ code files.

It is still possible to override this and do arbitrarily complicated
things for the tools that need this.
