+++ python/gradbench/gradbench/cpp_run.py
@@ -6,6 +6,7 @@ import sys
 import time
 from importlib import import_module
 from pathlib import Path
+import traceback
 
 
 def resolve(module, name):
@@ -23,12 +24,7 @@ def run(params):
         timings = list(map(json.loads, ls[1:]))
         return {"success": True, "output": output, "timings": timings}
     else:
-        return {
-            "success": False,
-            "status": proc.returncode,
-            "stderr": proc.stderr,
-            "stdout": proc.stdout,
-        }
+        return {"success": False, "status": proc.returncode, "error": proc.stderr}
 
 
 def main(pathname: str):
@@ -53,7 +49,7 @@ def main(pathname: str):
                     response["success"] = False
                 except Exception as e:
                     response["success"] = False
-                    response["error"] = str(e)
+                    response["error"] = traceback.format_exc() + "\n" + str(e)
             print(json.dumps({"id": message["id"]} | response), flush=True)
     except (EOFError, BrokenPipeError):
         pass
