--- tools/adol-c/AdolCBA.h
@@ -1,17 +1,0 @@
-#pragma once
-
-#include "adbench/shared/ITest.h"
-#include "adbench/shared/BAData.h"
-
-#include <vector>
-
-class AdolCBA : public ITest<BAInput, BAOutput> {
-private:
-  std::vector<double> _reproj_err_d;
-
-public:
-  AdolCBA(BAInput&);
-
-  virtual void calculate_objective() override;
-  virtual void calculate_jacobian() override;
-};
