+++ tools/adol-c/Makefile
@@ -1,12 +1,11 @@
 CXX?=c++
 CC?=cc
-CFLAGS=-std=c++17 -O3 -Wall
+CFLAGS=-std=c++17 -O3 -Wall -I../../cpp
 LDFLAGS=-lm -ladolc
 
 UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
 HELLO_OBJECTS=AdolCHello.o run_hello.o
 GMM_OBJECTS=AdolCGMM.o run_gmm.o
-BA_OBJECTS=AdolCBA.o run_ba.o
 HT_OBJECTS=AdolCHT.o run_ht.o
 LSTM_OBJECTS=AdolCLSTM.o run_lstm.o
 EXECUTABLES=run_hello run_gmm run_ba run_ht run_lstm
@@ -14,9 +13,6 @@ EXECUTABLES=run_hello run_gmm run_ba run_ht run_lstm
 all: $(EXECUTABLES)
 
 run_hello: $(HELLO_OBJECTS) $(UTIL_OBJECTS)
-	$(CXX) -o $@ $^ $(LDFLAGS)
-
-run_ba: $(BA_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
@@ -28,11 +24,11 @@ run_ht: $(HT_OBJECTS) $(UTIL_OBJECTS)
 run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
-%.o: %.c
-	$(CC) $< -c -o $@ $(CFLAGS) -I../../cpp
+run_%: run_%.cpp
+	$(CXX) -o $@ $^ $(LDFLAGS) $(CFLAGS)
 
 %.o: %.cpp
-	$(CXX) $< -c -o $@ $(CFLAGS) -I../../cpp
+	$(CXX) $< -c -o $@ $(CFLAGS)
 
 clean:
-	rm -f $(UTIL_OBJECTS) $(GMM_OBJECTS) $(BA_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
+	rm -f $(UTIL_OBJECTS) $(GMM_OBJECTS) $(HT_OBJECTS) $(EXECUTABLES)
