+++ tools/enzyme/Makefile
@@ -10,7 +10,7 @@ UTIL_OBJECTS=../../cpp/adbench/shared/utils.o ../../cpp/adbench/io.o
 HELLO_OBJECTS=EnzymeHello.o run_hello.o
 GMM_OBJECTS=EnzymeGMM.o run_gmm.o
 LSTM_OBJECTS=EnzymeLSTM.o run_lstm.o
-BA_OBJECTS=EnzymeBA.o run_ba.o
+BA_OBJECTS=run_ba.o
 HT_OBJECTS=EnzymeHT.o run_ht.o
 KMEANS_OBJECTS=run_kmeans.o
 EXECUTABLES=run_hello run_gmm run_lstm run_ba run_ht run_kmeans
@@ -26,7 +26,7 @@ run_gmm: $(GMM_OBJECTS) $(UTIL_OBJECTS)
 run_lstm: $(LSTM_OBJECTS) $(UTIL_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
-run_ba: $(BA_OBJECTS) $(UTIL_OBJECTS)
+run_ba: $(BA_OBJECTS)
 	$(CXX) -o $@ $^ $(LDFLAGS)
 
 run_ht: $(HT_OBJECTS) $(UTIL_OBJECTS)
